set toast

+++ src/Components/My Toys/ToysTable.jsx
@@ -1,21 +1,39 @@
 /* eslint-disable react/prop-types */
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
+
 const ToysTable = ({ toys, index, setToys, allToys }) => {
   const { name, email, img, price, description, photo, category,quantity, ratings, _id } = toys;
   const handledelete = (id) => {
     console.log(id)
-    fetch(`https://toys-market-server-site.vercel.app/delete/${id}`, {
-      method: "DELETE",
-      headers:{
-        "content-type":"application/json"
-      }
-    })
-      .then(res => res.json())
-      .then(data => {
-        console.log(data)
-        const remaining = allToys.filter(toy => toy._id !=id);
-        setToys(remaining)
-      })
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "Are you sure want to delete this ?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed){
+        fetch(`https://toys-market-server-site.vercel.app/delete/${id}`, {
+          method: "DELETE",
+          headers:{
+            "content-type":"application/json"
+          }
+        })
+          .then(res => res.json())
+          .then(data => {
+            console.log(data)
+            const remaining = allToys.filter(toy => toy._id !=id);
+            setToys(remaining)
+          })
+
+      }})
+
+    
+    
   }
   return (
     <tr>
